requirements.txt
Alfred-Workflow==1.18.2
requests-futures==0.9.7
iso8601==0.1.11
